Remove all of the Item and Relation classes
The Item and Relations were going to be a neat way of managing
Nepomuk-form relations in Baloo. The BalooCore library just contained
the abstract classes and the main implementation was the TagRelation
class which allowed one to have a separate tag database.

Overtime we have realized that Tags do not need a global database.
Specially since Akonadi will soon support server side tagging, and file
tags are being stored in the extended attributes.

Given that this "item and relation" classes are no longer being used for
anything, it might be better to not ship them right now. When we finally
have a use case for them then they can be added again.

This reduces Baloo to an abstract search provider along with File
metadata handling.