Support Akonadi instances in Baloo indexer
Use different Xapian databases for Baloo indexers running under
differents Akonadi instances. The default Xapian databases are
still stored in ~/.kde/share/apps/baloo/, but  when running under
a different Akonadi instance, the indexer will use databases in
~/.kde/share/apps/baloo/instances/$AKONADI_INSTANCE/. Applications
will also be able to search in these instance-specific databases
when AKONADI_INSTANCE env variable is set.

This patch is purely for comfort of a certain Akonadi developer,
who is using various Akonadi instances for testing and development
and does not want to have data from devel instances indexed in the
same database as data from his production instance. Applications
like KRunner will only be able to search in the default databases
(unless explicitely started with AKONADI_INSTANCE env set), but
that is something the certain Akonadi developer is willing to live
with. Akonadi instances are not used by normal users, so it will
not affect them.

REVIEW: 116708