Use XDG_DATA_HOME instead of KDE HOME
This is being done so that in the near future with KF5, the data
directory is kept the same. This will allow both KF5 Baloo and Qt4 Baloo
to easily co-exist and use the same database.

This will mean that everyone's database will need to be reindexed, but
since we haven't made a release so far. It shouldn't be a problem. Plus
indexing is quite fast.

REVIEW: 116764