SearchPluginTest: make the test stable
The "find by message by all status" testcase was flaky, because it was setting
conflicting flags on the item (Queued and Sent, and Watcher and Ignored) - during
conversion from Akonadi::Item::flags() to Akonadi::MessageStatus, setting Sent
unsets Queued, and setting Watched unsets Ignored (and vice versa). This was
broken with Qt5 where QHash is randomized so sometimes Watched went before
Ignored and sometimes Ignored before Watched, always unsetting the other.
The search query, which was trying to match the email by Ignored flag, then
failed in situation when Watched was after Ignored in the QSet, causing Ignored
to be unset and Watched set instead.