SearchPluginTest: only index the test data once, instead of for each test function
Presumably /tmp on the CI is not a tmpfs, so reindexing (and searching) takes
longer than on developer machines, causing the test to hit the 2 minutes timeout.
The search is read-only, so there's no need to re-index the data before every
test function. Only indexing the data once will save us some precious seconds
and hopefully getting us under the 2 minutes time limit on CI.