Create a new library for indexing and searching with Xapian
This removes most of the abstraction that we inherited from Baloo and only
leaves the minimal Xapian abstraction, since that's what we use. It puts
all indexing, searching and query parsing code into a single place instead
of having them spread around the Indexing Agent, Server plugin and Search
stores.
The stores and store plugins are now deprecated, the individual type support
is hardcoded into the library as there's no real reason to keep the type
support dynamic.

Next steps will be adding unit tests and porting the Indexing Agent and
Search plugin to this new library, then we can drop most of the original
code.