Add option to not build the KRunner plugin
We already disabled it on WIN32, now it is possible to disable it
manually at configure time via -DAKONADISEARCH_BUILD_KRUNNER_RUNNER=OFF.
The option is ON by default (except for WIN32).

This is useful for instance for Flatpak builds where we cannot integrate
with KRunner anyway, but it saves a lot of time by not having to build
KRunner, Plasma and KActivities Frameworks.